refactor(EditRolePlatforms): move reward adding logic to hook
